+++ src/main/java/com/proxiad/schultagebuch/repository/KlasseRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.proxiad.schultagebuch.entity.Klasse;
 
 @Repository
-public interface KlasseRepository extends JpaRepository<Klasse, Integer> {
+public interface KlasseRepository extends JpaRepository<Klasse, Long> {
 
 	public List<Klasse> findAllByOrderByIdAsc();
